gh-677: udpate url strucutre in imporeter and make it succeed
This can be seen as some sort of a spike to figure out the happy path.
While the importer succeeds it currently breaks all kinds of tests and
is in no way complete.

However, we pulled some good learnings out of this and now know what
parts are affected and what aren't.

We should consider squashing this branch in the end, so we don't end up
with individually breaking builds on our trunk.