gh-677: remove obsolete props from cref Model
This currently breaks in quite a few places because we don't seem to
have a good isolation between api constructs and internal constructs.
Most of the time this isolation happens through using schema datatypes
when talking to the db connectors, however in this case the
models.SingleRef seems to be widely used in both validation and even the
janus connector. This unfortunately leads to a massively breaking change
when changing the model. Next up we need to clean up the fallout.

The original reason for even starting this change was that acceptance
tests would fail with invalid uuid for $cref which is now indeed a URI,
rather than a UUID.