gh-677: add Ref type with options to parse from and stringify into
The intent is to clean up this models.SingleRef mess that is spread
throughout the validation and database packages. This new type should be
the only way to extract info about refs and can easily be parsed from
the API values and back into the API values to make sure we have an easy
way to use around the API without mixing up API design and business
logic.